Bind external properties to a dedicated class.
This allows to turn nullable `Boolean` fields from SentryOptions into a non-nullable `boolean` and keep simple getter and setter, without breaking Java beans standard.

Fixes #1738.
