+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -191,7 +191,7 @@ public final class Sentry {
 
   private static boolean initConfigurations(final @NotNull SentryOptions options) {
     if (options.isEnableExternalConfiguration()) {
-      options.merge(SentryOptions.from(PropertiesProviderFactory.create(), options.getLogger()));
+      options.merge(ExternalOptions.from(PropertiesProviderFactory.create(), options.getLogger()));
     }
 
     final String dsn = options.getDsn();
